% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_fun.R
\name{avail}
\alias{avail}
\title{What objects of this class are available}
\usage{
avail(classy, all_avail = TRUE)
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'Fleet')}

\item{all_avail}{Logical. If TRUE, function will return all objects of class \code{classy} available to user.
If FALSE, returns only those objects included in MSEtool.}
}
\description{
Generic class finder
}
\details{
Finds objects of the specified class in the global environment or in the
MSEtool and DLMtool packages. This function is an addendum to the \code{\link[DLMtool]{avail}}
function in DLMtool.
}
\examples{
avail("Assess")
avail("HCR")
avail("Stock")
avail("MP")
avail("MP", all_avail = FALSE)
}
\author{
Q. Huynh
}
