% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASAL2OM.R
\name{CASALpars}
\alias{CASALpars}
\title{Rips MLE estimates from CASAL file structure}
\usage{
CASALpars(CASALdir)
}
\arguments{
\item{CASALdir}{A folder with Stock Synthesis input and output files in it}
}
\value{
A list.
}
\description{
A function that uses the file location of a fitted CASAL assessment model including input files
to extract data required to populate an OMx class operating model.
}
\seealso{
\link{CASAL2OM}
}
\author{
T. Carruthers
}
