% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_auxiliary.R
\name{compare_models}
\alias{compare_models}
\title{Compare output from several assessment models}
\usage{
compare_models(..., label = NULL, color = NULL)
}
\arguments{
\item{...}{Objects of class \linkS4class{Assessment}.}

\item{label}{A character vector of the models for the legend.}

\item{color}{A vector of colors for each assessment model.}
}
\value{
A set of figures of biomass, recruitment, and fishing mortality estimates among the models.
}
\description{
Plot biomass, recruitment, and fishing mortality time series from several . This function can be used to compare outputs among
different assessment models from the same Data object.
}
\examples{
res <- cDD_SS(x = 3, Data = MSEtool::SimulatedData)
res2 <- SCA(x = 3, Data = MSEtool::SimulatedData)
res3 <- SP(x = 3, Data = MSEtool::SimulatedData)

compare_models(res, res2, res3)
}
\author{
Q. Huynh
}
