% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ancillary_Indicators.R
\name{mahplot}
\alias{mahplot}
\title{Plot statistical power of the indicator with increasing time blocks}
\usage{
mahplot(outlist, res = 6, maxups = 5, MPs)
}
\arguments{
\item{outlist}{A list object produced by the function \link{PRBcalc}}

\item{res}{Integer, the resolution (time blocking) for the calculation of PPD}

\item{maxups}{Integer, the maximum number of update time blocks to plot}

\item{MPs}{Character vector of MP names}
}
\description{
Plot statistical power of the indicator with increasing time blocks
}
\references{
Carruthers and Hordyk 2018
}
\author{
T. Carruthers
}
