% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_stock_multi_q_REDUNDANT.R
\name{optQ_multi}
\alias{optQ_multi}
\title{Optimize q for a single simulation}
\usage{
optQ_multi(
  par,
  depc,
  SSB0c,
  nareas,
  maxage,
  Ncurr,
  pyears,
  M_age,
  MatAge,
  Asize_c,
  WtAge,
  Vuln,
  Retc,
  Prec,
  movc,
  SRrelc,
  Effind,
  Spat_targc,
  hc,
  R0c,
  SSBpRc,
  aRc,
  bRc,
  maxF,
  MPAc,
  CFc,
  useCPP,
  mode = "opt"
)
}
\arguments{
\item{depc}{Depletion value}

\item{SSB0c}{Unfished spawning biomass}

\item{nareas}{Number of areas}

\item{maxage}{Maximum age}

\item{Ncurr}{Current N-at-age}

\item{pyears}{Number of years to project population dynamics}

\item{M_age}{M-at-age}

\item{MatAge}{Maturity-at-age}

\item{Asize_c}{Numeric vector (length nareas) with size of each area}

\item{WtAge}{Weight-at-age}

\item{Vuln}{Vulnerability-at-age}

\item{Retc}{Retention-at-age}

\item{Prec}{Recruitment error by year}

\item{movc}{movement matrix}

\item{SRrelc}{SR parameter}

\item{Effind}{Historical fishing effort}

\item{Spat_targc}{Spatial targetting}

\item{hc}{Steepness}

\item{R0c}{Unfished recruitment by area}

\item{SSBpRc}{Unfished spawning biomass per recruit by area}

\item{aRc}{Ricker aR}

\item{bRc}{Ricker bR}

\item{maxF}{maximum F}

\item{useCPP}{Logical. Use the CPP code?}

\item{logQ}{log q}

\item{MPA}{A matrix of spatial closures by year}

\item{model}{Character 'opt' is optimization and returns an objective funciton to be minimized, anything else returns fitted values}
}
\description{
Optimize q for a single simulation
}
\author{
A. Hordyk
}
\keyword{internal}
