% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{pcod}
\alias{pcod}
\title{Pacific cod in Area 5ABCD (Hecate Strait and Queen Charlotte Sound), British Columbia, Canada}
\format{
A list containing an object of class \linkS4class{OM}, \linkS4class{RCMdata}, and a list of priors
for index catchability.
}
\usage{
pcod
}
\description{
A list containing an operating model, data set, and priors for updating the operating model
using the conditioning model \link{RCM}.
}
\examples{
data(pcod)

}
\references{
Forrest, R.E., Anderson, S.C., Grandin, C.J., and Starr, P.J. 2020. Assessment of Pacific Cod (Gadus macrocephalus)
for Hecate Strait and Queen Charlotte Sound (Area 5ABCD), and West Coast Vancouver Island (Area 3CD) in 2018.
DFO Can. Sci. Advis. Sec. Res. Doc. 2020/070. v + 215 p.

DFO. 2021. Status Update of Pacifc Cod (Gadus macrocephalus) for West Coast Vancouver Island (Area 3CD),
and Hecate Strait and Queen Charlotte Sound (Area 5ABCD) in 2020. DFO Can. Sci. Advis. Sec. Sci. Resp. 2021/002.
}
\seealso{
\link{RCM}
}
\keyword{datasets}
