% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_S4_class_profile.R
\name{plot.prof}
\alias{plot.prof}
\alias{plot,prof,missing-method}
\title{Plot profile object}
\usage{
\S4method{plot}{prof,missing}(x, contour_levels = 20, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{prof} returned by \link{profile}.}

\item{contour_levels}{Integer, passed to \code{nlevels} argument of \link[graphics]{contour}.}

\item{...}{Miscellaneous. Not used.}
}
\description{
Generates a profile plot generated by \link{profile}. If a two-parameter profile is performed, then a contour
plot of the likelihood surface is returned.
}
\author{
Q. Huynh
}
