% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_S4_class_profile.R
\docType{class}
\name{prof-class}
\alias{prof-class}
\alias{prof}
\title{Class-\code{prof}}
\description{
An S4 class that contains output from \link{profile}.
}
\section{Slots}{

\describe{
\item{\code{Model}}{Name of the assessment model.}

\item{\code{Name}}{Name of Data object.}

\item{\code{Par}}{Character vector of parameters that were profiled.}

\item{\code{MLE}}{Numeric vector of the estimated values of the parameters (corresponding to \code{Par}) from the assessment.}

\item{\code{grid}}{A data.frame of the change in negative log-likelihood (\code{nll}) based on the profile of the parameters.}
}}

\seealso{
\link{plot.prof} \link{profile}
}
\author{
Q. Huynh
}
