% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_S4_class_project.R
\docType{class}
\name{project-class}
\alias{project-class}
\alias{project}
\title{Class-\code{project}}
\description{
An S4 class for the output from \link{projection}.
}
\section{Slots}{

\describe{
\item{\code{Model}}{Name of the assessment model.}

\item{\code{Name}}{Name of Data object.}

\item{\code{FMort}}{A matrix of fishing mortality over \code{p_sim} rows and \code{p_years} columns.}

\item{\code{B}}{An matrix of biomass with \code{p_sim} rows and \code{p_years} columns.}

\item{\code{SSB}}{A matrix of spawning biomass with \code{p_sim} rows and \code{p_years} columns.}

\item{\code{VB}}{A matrix of vulnerable biomass with \code{p_sim} rows and \code{p_years} columns.}

\item{\code{R}}{A matrix of recruitment over \code{p_sim} rows and \code{p_years} columns.}

\item{\code{N}}{A matrix of abundance over \code{p_sim} rows and \code{p_years} columns.}

\item{\code{Catch}}{A matrix of observed catch over \code{p_sim} rows and \code{p_years} columns.}

\item{\code{Index}}{A matrix of observed index over \code{p_sim} rows and \code{p_years} columns.}

\item{\code{C_at_age}}{An array for catch-at-age with dimension \code{c(p_sim, p_years, maxage)}.}
}}

\seealso{
\link{projection}
}
\author{
Q. Huynh
}
