% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCM_assess.R
\name{RCM_assess}
\alias{RCM_assess}
\title{The rapid conditioning model as an assessment function}
\usage{
RCM_assess(x = 1, Data, selectivity = c("logistic", "dome"), CAA_ESS = 50, ...)
}
\arguments{
\item{x}{A position in the Data object (by default, equal to one for assessments).}

\item{Data}{An object of class Data}

\item{selectivity}{Whether to model "logistic" or "dome" selectivity for the fishery.}

\item{CAA_ESS}{The annual multinomial sample size for the fishery age composition.}

\item{...}{Additional arguments (to be added).}
}
\description{
In beta testing. A function that uses RCM as an assessment function for use in MPs. Many more function arguments will be added
to tinker with model settings and data inputs. Currently uses catch, CAA, and biomass index from Data object, as well as biological
parameters in \code{Data@Misc$StockPars}.
}
\examples{
 
r <- RCM_assess(Data = SimulatedData)
myMP <- make_MP(RCM_assess, HCR_MSY)
myMP(x = 1, Data = SimulatedData)
}
