% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_S4_generic_simulate.R
\docType{class}
\name{sim-class}
\alias{sim-class}
\alias{sim}
\title{Class-\code{sim}}
\description{
An S4 class that contains output from \link{simulate}.
}
\section{Slots}{

\describe{
\item{\code{Model}}{Name of the assessment model.}

\item{\code{data}}{List of data from the assessment.}

\item{\code{data_sim}}{List of simulated data values. Each value returns an array.}

\item{\code{process_sim}}{List of simulated process error.}

\item{\code{est}}{Estimates from the original model fit.}

\item{\code{est_sim}}{Estimates from the simulated data.}
}}

\author{
Q. Huynh
}
