% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMSE_internal.R
\name{SOL}
\alias{SOL}
\title{Sum over list: get the list values from a list of lists}
\usage{
SOL(listy, namey)
}
\arguments{
\item{listy}{A list of objects}

\item{namey}{A character vector representing the list item's name}
}
\description{
Create of vector of values that correspond with a named position in a list of objects
}
\author{
T. Carruthers
}
