% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_S4_class_assess.R
\name{summary.Assessment}
\alias{summary.Assessment}
\alias{summary,Assessment-method}
\title{Summary of Assessment object}
\usage{
\S4method{summary}{Assessment}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{Assessment}}
}
\value{
A list of parameters.
}
\description{
Returns a summary of parameter estimates and output from an \linkS4class{Assessment} object.
}
\examples{
output <- DD_TMB(Data = DLMtool::Simulation_1)
summary(output)
}
