% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{swordfish}
\alias{swordfish}
\title{North Atlantic Swordfish dataset}
\format{
An object of class \linkS4class{Data}.
}
\source{
ASPIC Software at https://www.mhprager.com/aspic.html
}
\usage{
swordfish
}
\description{
An S4 object containing catch and index time series for North Atlantic swordfish.
}
\examples{
data(swordfish)

}
\keyword{datasets}
