% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOM_object.R
\name{tinyErr}
\alias{tinyErr}
\alias{tinyErr,MOM-method}
\title{Remove observation error, process error, implementation error or future gradients in (time varying) parameters}
\usage{
\S4method{tinyErr}{MOM}(x, obs = TRUE, imp = TRUE, proc = TRUE, grad = TRUE, silent = FALSE)
}
\arguments{
\item{x}{Object of class MOM. A Multi-OM object}

\item{obs}{Logical. Whether observation error should be removed}

\item{imp}{Logical. Whether implementation error should be removed}

\item{proc}{Logical. Whether process error should be removed (e.g. interannual variability in natural mortality rate)}

\item{grad}{Logical. Whether gradients (consistent temporal changes in parameters) should be removed}

\item{silent}{Logical. Should changes be printed to the console?}
}
\description{
This function allows like-by-like comparison of results among DLMtool `runMSE()` and MSEtool `multiMSE()`.
}
\author{
T.Carruthers and A. Hordyk
}
\keyword{internal}
