% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_MPs.R
\name{make_MP}
\alias{make_MP}
\title{Make a custom management procedure (MP)}
\usage{
make_MP(.Assess, .HCR, diagnostic = c("none", "min", "full"), ...)
}
\arguments{
\item{.Assess}{Assessment model, a function of class \code{Assess}.}

\item{.HCR}{Harvest control rule, a function of class \code{HCR}.}

\item{diagnostic}{A character string describing if any additional diagnostic information from the
assessment models will be collected during the closed-loop simulation (\code{"none"} is the
default). \code{"min"} (minimal) will collect information on convergence and \code{"full"} will also collect the
model estimates of biomass and F generated by \code{.Assess}.}

\item{...}{Additional arguments to be passed to \code{.Assess} and \code{.HCR}.}
}
\value{
A function of class \code{MP}.
}
\description{
Function operator that combines a function of class \code{Assess} and a function of
class \code{HCR} to create a management procedure (MP). The resulting function
can then be tested in closed-loop simulation via \code{\link[MSEtool]{runMSE}}.
}
\examples{
# A delay-difference model with a 40-10 control rule
DD_40_10 <- make_MP(DD_TMB, HCR40_10)

# A delay difference model that will produce convergence diagnostics
DD_40_10 <- make_MP(DD_TMB, HCR40_10, diagnostic = "min")

# MP that uses a Delay-Difference which assumes a Ricker stock-recruit function.
DD_Ricker <- make_MP(DD_TMB, HCR_MSY, SR = "Ricker")

\dontrun{
myMSE <- MSEtool::runMSE(MSEtool::testOM, MPs = c("FMSYref", "DD_40_10"))

diagnostic_AM(myMSE)
}

}
\seealso{
\link{HCR_ramp} \link{HCR_MSY} \link{diagnostic_AM} \link{retrospective_AM}
}
