% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_fun.R
\name{userguide}
\alias{userguide}
\title{Get the SAMtool vignettes}
\usage{
userguide()
}
\value{
Displays a browser webpage to the openMSE website.
}
\description{
A convenient function to open a web browser with the openMSE documentation vignettes
}
\examples{
userguide()

}
