% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_fun.R
\name{userguide}
\alias{userguide}
\title{Get the SAMtool vignettes}
\usage{
userguide()
}
\description{
A convenient function to open a web browser with the SAMtool package vignettes
}
\examples{
\dontrun{
userguide()
}
}
\seealso{
\link[OMtool]{userguide}
}
