% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_fun.R
\name{userguide}
\alias{userguide}
\title{Get the MSEtool vignettes}
\usage{
userguide()
}
\description{
A convenient function to open a web browser with the MSEtool package vignettes
}
\examples{
\dontrun{
MSEtool::userguide()
DLMtool::userguide()
}
}
\seealso{
\link[DLMtool]{userguide}
}
